move the EOA estimation to a different module
